requirements.txt
streamlit
google-generativeai
python-dotenvpi